Clear the rest of the reliability findings (#62)
Follows #61. Between them these take SonarCloud's reliability count from
35 to 5.

WXPermissions.checkWXPermissions opened with a null guard on the player.
No caller can reach it with null -- all thirty-nine pass one from an
instanceof cast, a Bukkit event, or a local derived from one -- so it
never fired. It was not harmless: Sonar reads it as proof the parameter is
nullable, and since every caller messages that same player in the else
branch, seven javabugs:S2259 reports across Build, Compass, Dial,
RingCommand, BeamCommand and GateInteractionHandler traced back to this
one line. Deleting it clears all seven, with no behaviour change because
the branch was unreachable. Same shape as the StargateUpdateRunnable fix
in #61: the fault was the thing permitting an impossible state, not the
dereferences that state implied.

removeStargate ignored the result of delete(), so a gate file that could
not be removed came back on the next load in silence; the config writer
ignored createNewFile the same way. Both report now.

Ten S2184 sites computed block coordinates in int and let the result
widen. None can overflow at Minecraft's coordinate limits, so the casts
are tidying rather than a fix.

The first version of the delete warning called getThisPlugin() unguarded,
three lines below a getGatesDir that explicitly tolerates it being null.
Sonar caught that as a new blocker on the PR while every required check
was green. Fixed before merge; the delete still happens either way and
only the report is conditional.

Left alone: five java:S3077 volatile-collection fields, which are a real
concurrency decision rather than a mechanical change.

754 tests pass. Full CI matrix green; SonarCloud quality gate passes.

+++ src/main/java/com/wormhole_xtreme/wormhole/model/Stargate.java
@@ -421,7 +421,7 @@ public class Stargate
         // Block centres, so a traveller lands in the middle of the block rather than on its
         // corner, and at its floor rather than inside it.
         final double x = Math.round((double) sumX / counted) + 0.5 + stepX;
-        final double y = lowest + stepY;
+        final double y = (double) lowest + stepY;
         final double z = Math.round((double) sumZ / counted) + 0.5 + stepZ;
         final Location fixed = new Location(gateWorld, x, y, z);
         try
